requirements.txt
# IRIS Economic System Simulation - Dependencies
# ===============================================

# Core scientific computing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0

# Visualization
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0  # For interactive visualizations

# Configuration
pyyaml>=6.0,<7.0.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-timeout>=2.1.0,<3.0.0

# Code quality
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.0.0,<2.0.0

# Progress bars and CLI
tqdm>=4.65.0,<5.0.0

# Data export (optional but recommended)
openpyxl>=3.0.0,<4.0.0  # For Excel export

# Memory profiling (optional)
memory_profiler>=0.61.0,<1.0.0
psutil>=5.9.0,<6.0.0

# For notebook support (optional)
# jupyter>=1.0.0
# ipython>=7.0.0
